refactor: rename key_exists.rs to not conflict with submodule
We cannot have a sub-module with the smae name as the file. Since we
want to group both `key_exists` and `keys_exists` in the same module,
we change the name of the file.
