+++ tests/cache/key_existence.rs
